pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "beautifulsoup4>=4.14.2",
    "celery>=5.5.3",
    "eventlet>=0.40.3",
    "flask-sock>=0.7.0",
    "flask-sqlalchemy>=3.1.1",
    "lxml>=6.0.2",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
]
